fix pageout waiting time and deleting cgroup dir for e2e tests

+++ test/e2e/memtierd.test-suite/memtierd.source.sh
@@ -110,7 +110,8 @@ memtierd-meme-start() {
 
 memtierd-meme-stop() {
     vm-command "killall -KILL meme"
-    if [[ -n "$MEME_CGROUP" ]]; then
+    sleep 2
+    if [[ -z "$MEME_PID" ]] && [[ -n "$MEME_CGROUP" ]]; then
         vm-command "sudo rmdir /sys/fs/cgroup/$MEME_CGROUP"
     fi
 }
